Conflicts:
	loaders/CSharp/GeneratorUnitTests/GetCodegenTypeTests.cs

+++ loaders/CSharp/GeneratorLib/CodeGenerator_GenerateCode.cs
@@ -114,7 +114,7 @@ namespace GeneratorLib
                 {
                     if (schema.Default != null)
                     {
-                        defaultValue = new CodePrimitiveExpression((float)schema.Default);
+                        defaultValue = new CodePrimitiveExpression((float)(double)schema.Default);
                     }
                     return new CodeTypeReference(typeof(float));
                 }
